Moved backtrace code into a separate file
